client: Do not remove the provided `--update-file`
If you provide a local update JSON file, it is unexpected for the client
to actually delete it at the end of the query procedure.

+++ steamosatomupd/client.py
@@ -812,7 +812,8 @@ class UpdateClient:
         if args.query_only:
             if not args.quiet:
                 print(json.dumps(update.to_dict(), indent=2))
-            os.remove(update_file)
+            if not args.update_file:
+                os.remove(update_file)
             return 0
 
         # Apply update
